Move DatabaseExtras back to trace Add a new BlockChainWithExtras newtype to ethcore Impl DatabaseExtras for BlockChainWithExtras
